few endpoint small fixes

+++ back/User/user_router.py
@@ -1,11 +1,13 @@
 from datetime import datetime, timedelta, timezone
 from typing import List
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import StreamingResponse
 import pandas as pd
 import io
 import openpyxl
 
+from back.User.depends import get_current_user
+from back.User.models import User
 from back.User.schemas import SUser, SUserCreate, SUserAllSchema
 from back.User import crud
 
@@ -16,7 +18,11 @@ router = APIRouter(
 
 email_regex = r"^[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$"
 
-@router.get("users/to_excel_file")
+@router.get("/{user_id}/fullname", response_model=str)
+async def get_user_full_name(user_id: int):
+    return await crud.get_user_full_name(user_id)
+
+@router.get("/to_excel_file")
 async def get_software_excel():
     user_list = await crud.get_all_users()
 
@@ -52,11 +58,17 @@ async def get_software_excel():
         headers={"Content-Disposition": f"attachment; filename={file_name}"}
     )
     
-@router.get("users/all")
+@router.get("/all")
 async def get_all_users() -> List[SUserAllSchema]:
-    
     return await crud.get_all_users()
 
-@router.delete("/{user_id}", response_model=dict)
-async def delete_user(user_id: int):
-    return await crud.delete_user(user_id=user_id)
+@router.delete("/{user_id}")
+async def delete_user(user_id: int, user: User = Depends(get_current_user)):
+    if user.system_role_id == 4:
+        return await crud.delete_user(user_id=user_id)
+    else: 
+        return HTTPException(status_code=403, detail="Forbidden")
+    
+@router.put("/{user_id}", response_model=SUser)
+async def update_user(updated_user: SUser):
+    return await crud.update_user(updated_user)
